Break up models module
