+++ project/migrate.py
@@ -9,8 +9,8 @@ from textwrap import dedent
 import pandas as pd
 from pydantic import BaseModel, DirectoryPath, Field
 
-from boilerdata.common import StrPath, dump_model, load_config, write_schema
-from boilerdata.models import Project, get_project
+from boilerdata.models.common import StrPath, dump_model, load_config, write_schema
+from boilerdata.models.project import Project, get_project
 
 
 def main():
