+++ project/try_00_act_vs_est.py
@@ -45,7 +45,7 @@ import pandas as pd
 from scipy.stats import linregress, t
 
 from boilerdata.axes import Axes as A  # noqa: N817
-from boilerdata.models import get_project
+from boilerdata.models.project import get_project
 from boilerdata.pipeline import get_df
